+++ src/main/java/chat/woowa/woowachat/chat/domain/Messages.java
@@ -1,4 +1,4 @@
-package chat.woowa.woowachat.chat;
+package chat.woowa.woowachat.chat.domain;
 
 import static jakarta.persistence.CascadeType.ALL;
 import static jakarta.persistence.FetchType.LAZY;
