+++ src/main/java/chat/woowa/woowachat/chat/domain/SettingMessage.java
@@ -1,4 +1,4 @@
-package chat.woowa.woowachat.chat;
+package chat.woowa.woowachat.chat.domain;
 
 public enum SettingMessage {
