+++ src/main/java/chat/woowa/woowachat/member/domain/Course.java
@@ -1,4 +1,4 @@
-package chat.woowa.woowachat.member;
+package chat.woowa.woowachat.member.domain;
 
 public enum Course {
