+++ src/main/java/chat/woowa/woowachat/member/domain/MemberRepository.java
@@ -1,7 +1,8 @@
-package chat.woowa.woowachat.member;
+package chat.woowa.woowachat.member.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
+    boolean existsByName(final String name);
 }
