+++ src/test/java/chat/woowa/woowachat/chat/fixture/MessageFixture.java
@@ -1,7 +1,7 @@
 package chat.woowa.woowachat.chat.fixture;
 
-import chat.woowa.woowachat.chat.Message;
-import chat.woowa.woowachat.chat.Messages;
+import chat.woowa.woowachat.chat.domain.Message;
+import chat.woowa.woowachat.chat.domain.Messages;
 import java.util.Arrays;
 import java.util.List;
